requirements.txt
soundfile
numpy
torch
torchvision
torchaudio
tokenizers
encodec
langid
wget
unidecode
pyopenjtalk
pypinyin
inflect
cn2an
jieba
eng_to_ipa
openai-whisper
matplotlib
tensorboard
gradio==3.41.2
nltk
sudachipy
sudachidict_core
vocos
h5py
torchmetrics
peft
gdown
librosa
pyworld
praat-parselmouth
language-tool-python
pesq
jiwer